% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{AUGglmmTMB}
\alias{AUGglmmTMB}
\title{Fit a penalized binomial mixed model}
\usage{
AUGglmmTMB(
  formula,
  data,
  weights = NULL,
  link = "logit",
  penOpt = AUGglmmTMBPenalty(),
  control = AUGglmmTMBControl()
)
}
\arguments{
\item{formula}{A \code{\link{formula}} object specifying the fixed and random
effects in standard R formula syntax. Random effects should be of the form
\code{(expr | factor)}.}

\item{data}{A \code{data.frame} containing all variables used in \code{formula}. When missing values are present, the observations with any missing value are removed as in \code{\link{na.omit}}.}

\item{weights}{Optional. Column name, as in \code{\link{glmmTMB}} in \code{data} specifying
observation weights. Default is \code{NULL}, in which case all weights are 1.}

\item{link}{Character. Link function to use for the binomial response. Default is
\code{"logit"}.}

\item{penOpt}{List of penalty options, created by \code{\link{AUGglmmTMBPenalty}}:
\describe{
  \item{\code{cfe}}{Numeric. Strength of penalty on fixed effects. If \code{NULL} (the default),
    internally computed via \code{\link{mv_multiplier}} as suggested by Košuta et al. Setting \code{cfe = 0} disables fixed-effect penalty.}
  \item{\code{autrepen}}{Logical. If \code{TRUE}, automatically estimates the random-effects
    penalty parameters \eqn{\tau} and \eqn{\Psi} using the procedure described in Košuta et al (see also \code{\link{get_psi}}). When \code{TRUE}, the penalty is only applied to the 1st random effect. Default is \code{FALSE}.}
  \item{\code{nu}}{Optional list specifying random-effects penalty parameters. Ignored when \code{autrepen=TRUE}. Can be of the same length as the number of specified random effects in which case the penalty is applied to all random effects; when shorter, the penalty is applied only to the first \code{length(nu)} random effects. Default is \code{NULL} in which case the random effects are not penalized.}
  \item{\code{psi}}{Optional list of random-effects penalty matrices. Ignored when \code{autrepen=TRUE}. Ignored when \code{nu=NULL}; when \code{nu} is not \code{NULL}, \code{psi} has to be the list of the same length as \code{nu} containing matrices of appropriate dimensions. Default is \code{NULL}.}
 \item{\code{tau}}{Optional list of penalty parameters. Ignored when \code{autrepen=TRUE} and when \code{nu} is not null. When specified, this has to be the list of numeric values in [0,1]; the list can be of the same length as the number of specified random effects in which case the penalty is applied to all random effects; when shorter, the penalty is applied only to the first \code{length(tau)} random effects. Default is \code{NULL}.}
\item{\code{plot}}{Logical; when \code{TRUE} the conditional likelihood by \eqn{\tau} is plotted. Ignored when \code{autrepen=FALSE}. When \code{TRUE}, the computation time can be substantially increased. Default is \code{FALSE}.}
\item{\code{ntaus}}{Numeric, number of different values of \eqn{\tau} to be used when plotting the conditional likelihood. Ignored when \code{plot=FALSE}. Using a large value can substantially increase the computation time. Default is \code{50}.}
}}

\item{control}{List of control parameters created via \code{\link{AUGglmmTMBControl}}:
\describe{
\item{\code{fit_pGLM}}{Logical. If \code{TRUE}, uses the penalized GLM step (Algorithm 2
  in Košuta et al.), otherwise it uses Algorithm 1. Default is \code{FALSE}.}
\item{\code{method_pGLM}}{Character. The method to be used when fitting penalized GLM.
Currently supported values are \code{"brglm2"} and \code{"logistf"}.
\code{"logistf"} can only be used for Bernoulli outcomes when \code{link_fun="logit"}.
Ignored when \code{fit_pGLM=FALSE}. Defaults to \code{"brglm2"}.}
\item{\code{maxiter}}{Integer. Maximum number of iterations for the penalized likelihood
  fitting procedure. Default is 50.}
\item{\code{tol}}{Numeric. Convergence tolerance based on changes in parameter estimates.
  Default is 1e-6.}
\item{\code{save_coef}}{Logical. If \code{TRUE}, stores coefficient estimates at each
  iteration. Default is \code{FALSE}.}
\item{\code{inter_iter}}{Integer. Number of internal iterations used by
  \pkg{glmmTMB}. Default is 1000.}
\item{\code{use_previous}}{Logical. If \code{TRUE}, uses estimates from the previous
  iteration as starting values. Default is \code{FALSE}.}
}}
}
\value{
A list with elements:
\describe{
  \item{fit}{The fitted penalized GLMM object of class \code{\link{glmmTMB}}.}
  \item{loglik}{The unpenalized marginal log-likelihood.}
  \item{coefs}{If \code{save_coef=TRUE}, a list with elements \code{coefs_glmm} and \code{coefs_glm}
  (corresponding to the matrices of GLMM and GLM parameters by iteration, respectively), otherwise \code{NULL}.}
  \item{optre}{If \code{autrepen = TRUE}, a list with elements \code{opt_tau} and \code{opt_psi}
  (corresponding to the estimated penalty parameter \eqn{\tau} and penalty matrix \eqn{\Psi}, respectively), otherwise \code{NULL}.}
}
}
\description{
Fits a penalized binomial generalized linear mixed model with optional
penalties on both fixed and random effects using the approach of Košuta et al.
}
\details{
Setting \code{autrepen=TRUE} uses the data-driven procedure proposed by Košuta et al. to determine the penalty parameters; the parameter \code{nu} is set to \eqn{2q-1} internally, any other value supplied in \code{nu} is ignored.
If \code{autrepen=FALSE}, \code{nu=NULL}, and \code{tau=NULLL}, no random-effects penalty is applied and only the fixed-effects penalty is used.
When \code{autrepen=FALSE}, if provided, \code{nu} and \code{psi}, can be of the same length as the number of specified random effects, in which case
the penalty is applied to all random effects. If shorter, the penalty is applied
only to the first \code{length(nu)} random-effects terms. When specified, \code{nu} and \code{psi} need to be of the same length: each element of \code{nu} and \code{psi} are the penalty parameters \eqn{\nu} and \eqn{\Psi}, respectively, for the corresponding random effect. When specified, the elements of \code{psi} need to be matrices of appropriate dimensions.
When \code{autrepen=FALSE} and \code{nu=NULL}, the penalty parameters of the random effects, can also be specified via the argument \code{tau}. This has to be a list, that can be of the same length as the number of specified random effects, in which case
the penalty is applied to all random effects. If shorter, the penalty is applied
only to the first \code{length(tau)} random-effects terms. When specified, each element of the list has to be a numeric value in [0,1]. For each element in the list, its respective random effects parameters are set to \eqn{2q-1} and \eqn{\bm{\Psi} = 3q \bm{T} \tilde{\bm{\Lambda}}_\tau \bm{T}^\top}, where \eqn{\bm{\bm\Sigma} = \bm{T} \bm{\Lambda} \bm{T}^\top}, \eqn{\bm{T} \bm{T}^\top = \bm{I}_q} and \eqn{\bm{\Lambda} = \operatorname{diag}(\lambda_1, \dots, \lambda_q)} with \eqn{\lambda_1 \geq \cdots \geq \lambda_q} being the eigenvalues of \eqn{\bm{\bm\Sigma}} and \eqn{\tilde{\bm{\Lambda}}(\tau)=\tilde{\bm{\Lambda}}_\tau} is a diagonal matrix with diagonal elements \eqn{\tilde{\lambda}_i = \lambda_i + \tau(1 - \lambda_i)}, \eqn{i=1,\ldots,q}, where \eqn{\tau \in [0,1]} is the shrinkage parameter set in \code{tau}. The decomposition is based on the estimate of \eqn{\bm\Sigma} obtained when only the fixed-effects penalty is applied.
}
\examples{
\dontrun{
data(birds)

#default fixed-effects penalty, no penalty on the random effect;
#single-iteration approximation to MPL

fit1 <- AUGglmmTMB(
  cbind(parasites, 1 - parasites) ~ migration + food + (1 | species),
  data = birds,
  weights = NULL,
  link = "logit",
  penOpt = AUGglmmTMBPenalty(),
  control = AUGglmmTMBControl(fit_pGLM = TRUE, maxiter = 1)
)
summary(fit1$fit)

#the same

fit2 <- AUGglmmTMB(
  parasites ~ migration + food + (1 | species),
  data = birds,
  weights = NULL,
  link = "logit",
  penOpt = AUGglmmTMBPenalty(),
  control = AUGglmmTMBControl(fit_pGLM = TRUE, maxiter = 1)
)
summary(fit2$fit)

#default penalty on the fixed effects, penalty on both random effects;
#single-iteration approximation to MPL

fit3 <- AUGglmmTMB(
  parasites ~ migration + food + (migration|phylogenetic) + (1 | species),
  data = birds,
  weights = NULL,
  link = "logit",
  penOpt = AUGglmmTMBPenalty(nu=list(3,1),psi=list(diag(1,2,2),matrix(1,1,1))),
  control = AUGglmmTMBControl(fit_pGLM = TRUE, maxiter = 1)
)
summary(fit3$fit)

#default penalty on the fixed effects, data-driven penalty on phylogenetic, no penalty on species;
#single-iteration approximation to MPL

fit4 <- AUGglmmTMB(
  parasites ~ migration + food + (migration|phylogenetic) + (1 | species),
  data = birds,
  weights = NULL,
  link = "logit",
  penOpt = AUGglmmTMBPenalty(autrepen=TRUE,plot=TRUE,ntaus =20),
  control = AUGglmmTMBControl(fit_pGLM = TRUE, maxiter = 1)
)

fit4$optre
summary(fit4$fit)
}

}
\seealso{
\code{\link{AUGglmmTMBPenalty}},\code{\link{AUGglmmTMBControl}}, \code{\link{get_psi}}, \code{\link{mpl_fitter}}
}
