% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{AUGglmmTMBPenalty}
\alias{AUGglmmTMBPenalty}
\title{Create penalty options for AUGglmmTMB}
\usage{
AUGglmmTMBPenalty(
  cfe = NULL,
  autrepen = FALSE,
  nu = NULL,
  psi = NULL,
  tau = NULL,
  plot = FALSE,
  ntaus = 50
)
}
\arguments{
\item{cfe}{Numeric. Strength of penalty on fixed effects. Default is \code{NULL},
in which case \code{AUGglmmTMB} computes it internally via \code{\link{mv_multiplier}} as suggested by Košuta et al.
Setting \code{cfe = 0} disables the fixed-effects penalty.}

\item{autrepen}{Logical. If \code{TRUE}, automatically estimates the random-effects
penalty parameters \eqn{\tau} and \eqn{\Psi} using the procedure described in Košuta et al (see also \code{\link{get_psi}}). Default is \code{FALSE}.
When \code{TRUE}, the penalty is only applied to the 1st random effect.}

\item{nu}{Optional list of numeric vectors specifying random-effects penalty parameters (degrees of freedom). If not \code{NULL}, \code{nu} must be a list in which each element is a single numeric value. Can be of the same length as the number of specified random effects in which case the penalty is applied to all random effects; when shorter, the penalty is applied only to the first \code{length(nu)} random effects. Ignored when \code{autrepen=TRUE}. Default is \code{NULL} in which case the random effects are not penalized.}

\item{psi}{Optional list of random-effects penalty matrices. Ignored when \code{autrepen=TRUE}. Ignored when \code{nu=NULL}; when \code{nu} is not \code{NULL}, \code{psi} has to be the list of the same length as \code{nu} containing matrices of appropriate dimensions. Default is \code{NULL}.}

\item{tau}{Optional list of penalty parameters. Ignored when \code{autrepen=TRUE} and when \code{nu} is not null. When specified, this has to be the list of numeric values in [0,1]; the list can be of the same length as the number of specified random effects in which case the penalty is applied to all random effects; when shorter, the penalty is applied only to the first \code{length(tau)} random effects. Default is \code{NULL}.}

\item{plot}{Logical; when \code{TRUE} the conditional likelihood by \eqn{\tau} is plotted. Ignored when \code{autrepen=FALSE}. When \code{TRUE}, the computation time can be substantially increased. Default is \code{FALSE}.}

\item{ntaus}{Numeric, number of different values of \eqn{\tau} to be used when plotting the conditional likelihood. Ignored when \code{plot=FALSE}. Using a large value can substantially increase the computation time. Default is \code{50}.}
}
\value{
A named list containing all specified penalty parameters.
}
\description{
Constructs a list of penalty parameters to be used with \code{\link{AUGglmmTMB}}.
}
\details{
This function does not perform any fitting itself. It simply returns a structured list
that specifies how \code{\link{AUGglmmTMB}} should apply penalties to fixed and random effects.
Use this function to easily create the \code{penOpt} argument for \code{\link{AUGglmmTMB}}.
Setting \code{autrepen=FALSE}, \code{nu=NULL}, and \code{tau=NULL} turns off the penalty on the random effects covariance matrices.
}
\examples{
# Default penalty on the fixed effects, no penalty on the random effects
pen <- AUGglmmTMBPenalty()

# Fixed-effects penalty strength set to 0.5, data-driven random-effects penalty parameters
pen2 <- AUGglmmTMBPenalty(cfe = 0.5, autrepen = TRUE)

# No fixed-effects penalty, data-driven random-effects penalty parameters
pen3 <- AUGglmmTMBPenalty(cfe = 0, autrepen = TRUE)

# Turn off both penalties (the same as glmmTMB)
pen4 <- AUGglmmTMBPenalty(cfe = 0)

}
