% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{get_psi}
\alias{get_psi}
\title{Estimate penalty parameters for random effects}
\usage{
get_psi(
  data,
  cfe,
  nu,
  fit_pGLM = FALSE,
  maxiter = 50,
  tol = 1e-06,
  link_fun = "logit",
  save_coef = FALSE,
  inter_iter = 1000,
  use_previous = FALSE
)
}
\arguments{
\item{data}{A list containing the model data. See \code{mpl_fitter} for details.}

\item{cfe}{Numeric. Strength of the penalty applied to the fixed effects.}

\item{nu}{Penalty parameter for the random effects. See \code{mpl_fitter}
for details.}

\item{fit_pGLM}{Logical. Indicates whether Algorithm 2 (\code{TRUE}) or Algorithm 1 (\code{FALSE}) is used. Default is \code{FALSE}.}

\item{maxiter}{Integer. Maximum number of iterations. Default is \code{50}.}

\item{tol}{Numeric. Convergence tolerance. Default is \code{1e-6}.}

\item{link_fun}{Character. Link function. Default is \code{"logit"}.}

\item{save_coef}{Logical. Whether to store coefficient paths. Default is \code{FALSE}.}

\item{inter_iter}{Integer. Number of internal iterations used by \pkg{glmmTMB}. Default is \code{1e3}.}

\item{use_previous}{Logical. Whether to use previous estimates as starting values. Default is \code{TRUE}.}
}
\value{
A list with elements:
\describe{
  \item{tau}{Estimated value of \eqn{\tau}.}
  \item{psi}{Estimated value of \eqn{\Psi}.}
}
}
\description{
Computes the penalty parameters \eqn{\tau} and \eqn{\Psi} for the random-effects
penalty, following the approach of Košuta et al. The penalty is assumed to be
applied only to the first random-effects term.
}
\details{
This function is used internally by \code{\link{AUGglmmTMB}} and is not meant to be used directly (it can be, if used correctly, however (much) faster).

The procedure assumes that the random-effects penalty is imposed only on the
first random-effects component. The estimates are obtained using an iterative
procedure consistent with the penalized likelihood framework described in
Košuta et al.
}
\examples{
\dontrun{

data(birds)
Y<-birds$parasites
X <- model.matrix(parasites~migration+food, birds)
Z1<-model.matrix(parasites~migration, birds)
Z2<-model.matrix(parasites~1, birds)
grouping1<-as.numeric(as.factor(birds$Phylogenetic.Tomi))
grouping2<-as.numeric(as.factor(birds$species))
xdf<-list(Y=Y,X=X,Z=list(Z1,Z2),
  grouping=list(grouping1,grouping2),
  M=rep(1,nrow(X)),W=rep(1,nrow(X)))

psi_opt<-get_psi(xdf,mv_multiplier(xdf),nu=list(3),
  fit_pGLM=TRUE,maxiter=1,tol=1e-6,link_fun="logit",
    save_coef=FALSE,inter_iter=1e3,use_previous=FALSE)
psi_opt

}
}
\seealso{
\code{\link{AUGglmmTMB}}, \code{\link{get_data_plot_cloglik}}, \code{\link{mpl_fitter}}
}
