% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{mpl_fitter}
\alias{mpl_fitter}
\title{Fit a penalized binomial mixed model via iterative algorithms}
\usage{
mpl_fitter(
  data,
  cfe,
  nu,
  psi,
  fit_pGLM = FALSE,
  maxiter = 50,
  tol = 1e-06,
  link_fun = "logit",
  save_coef = FALSE,
  inter_iter = 1000,
  use_previous = TRUE
)
}
\arguments{
\item{data}{A list containing the model data with elements:
\describe{
  \item{Y}{Response vector.}
  \item{M}{Number of trials. Can be missing in which case it is internally set to 1 for all observations.}
  \item{W}{Likelihood, frequency or precision weights. Can be missing in which case it is internally set to 1 for all observations.}
  \item{X}{Design matrix for fixed effects.}
  \item{Z}{Random-effects design matrix or a list of matrices for multiple
  random-effects terms.}
  \item{grouping}{Grouping variable or a list of grouping variables corresponding
  to each random-effects term.}
}
If multiple random effects are present, \code{Z} and \code{grouping} must be
lists of equal length.}

\item{cfe}{Numeric. Strength of the penalty applied to the fixed effects
(as in Košuta et al.).}

\item{nu, psi}{Penalty parameters for the random effects. If \code{NULL},
no random-effects penalty is applied and only the fixed-effects penalty is used.
If provided, these have to be lists, that can be of the same length as \code{Z}, in which case
the penalty is applied to all random effects. If shorter, the penalty is applied
only to the first \code{length(nu)} random-effects terms.}

\item{fit_pGLM}{Logical. If \code{TRUE}, uses Algorithm 2; if \code{FALSE}, uses Algorithm 1. Default is \code{FALSE}.}

\item{maxiter}{Integer. Maximum number of iterations. Default is \code{50}}

\item{tol}{Numeric. Convergence tolerance based on changes in the parameter
vector (fixed effects and variance components). Default is \code{1e-6}.}

\item{link_fun}{Character. Link function to use. Supported options include
\code{"logit"}, \code{"probit"}, \code{"cloglog"}, \code{"cauchit"}, and
\code{"loglog"}. Note that \code{"cauchit"} and \code{"loglog"} are not
available in \pkg{glmmTMB}, although they are supported internally. Default is \code{"logit"}.}

\item{save_coef}{Logical. If \code{TRUE}, stores the parameter estimates
(fixed effects and variance components) at each iteration. Default is \code{FALSE}.}

\item{inter_iter}{Integer. Number of internal iterations used by
\pkg{glmmTMB}. Default is \code{1e3}.}

\item{use_previous}{Logical. If \code{TRUE}, uses parameter estimates from the
previous iteration as starting values. Default is \code{TRUE}.}
}
\value{
A list with the following elements:
\describe{
  \item{fit}{The fitted penalized GLMM object.}
  \item{loglik}{The unpenalized marginal log-likelihood.}
  \item{coefs}{Matrix of GLMM parameter estimates by iteration.}
  \item{coefs_glm}{Matrix of GLM parameter estimates by iteration (only relevant
  when \code{fit_pGLM = TRUE}).}
}
}
\description{
Fits a penalized binomial mixed model (GLMM) using Algorithm 1 or Algorithm 2 presented in Košuta et al.
}
\details{
The function implements penalized likelihood estimation for binomial mixed model with optional
penalties on both fixed and random effects. Two algorithms are available:
Algorithm 1 and Algorithm 2. The single-iteration estimator from Košuta et el. is obtained by specifying \code{fit_pGLM = TRUE,maxiter=1}.
}
\examples{
\dontrun{


data(birds)
Y<-birds$parasites
X <- model.matrix(parasites~migration+food, birds)
Z1<-model.matrix(parasites~migration, birds)
Z2<-model.matrix(parasites~1, birds)
grouping1<-as.numeric(as.factor(birds$Phylogenetic.Tomi))
grouping2<-as.numeric(as.factor(birds$species))
xdf<-list(Y=Y,X=X,Z=list(Z1,Z2),
  grouping=list(grouping1,grouping2),
  M=rep(1,nrow(X)),W=rep(1,nrow(X)))

#MPL-A1

fit_mpl1<-mpl_fitter(xdf,mv_multiplier(xdf),nu=list(3),psi=list(diag(1,2,2)),
  fit_pGLM=FALSE,maxiter=50,tol=1e-6,link_fun="logit",
    save_coef=TRUE,inter_iter=1e3,use_previous=TRUE)
summary(fit_mpl1$fit)
fit_mpl1$coefs

#MPL-A2

fit_mpl2<-mpl_fitter(xdf,mv_multiplier(xdf),nu=list(3),psi=list(diag(1,2,2)),
  fit_pGLM=TRUE,maxiter=50,tol=1e-6,link_fun="logit",
    save_coef=TRUE,inter_iter=1e3,use_previous=TRUE)
summary(fit_mpl2$fit)
fit_mpl2$coefs

#' #MPL-A2(1)

fit_mpl21<-mpl_fitter(xdf,mv_multiplier(xdf),nu=list(3),psi=list(diag(1,2,2)),
  fit_pGLM=TRUE,maxiter=1,tol=1e-6,link_fun="logit",
    save_coef=TRUE,inter_iter=1e3,use_previous=TRUE)
summary(fit_mpl21$fit)
fit_mpl21$coefs

#MPL-A1: penalty on bith REs

fit_mpl1_2repen<-mpl_fitter(xdf,mv_multiplier(xdf),nu=list(3,1),psi=list(diag(1,2,2),diag(1,1,1)),
  fit_pGLM=FALSE,maxiter=50,tol=1e-6,link_fun="logit",
    save_coef=TRUE,inter_iter=1e3,use_previous=TRUE)
summary(fit_mpl1_2repen$fit)
fit_mpl1_2repen$coefs
}
}
\seealso{
\code{\link{get_psi}}, \code{\link{get_data_plot_cloglik}}
}
