% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{AUGglmmTMB}
\alias{AUGglmmTMB}
\title{Fit a penalized binomial mixed model}
\usage{
AUGglmmTMB(
  formula,
  data,
  weights = NULL,
  link = NULL,
  penOpt = AUGglmmTMBPenalty(),
  control = AUGglmmTMBControl()
)
}
\arguments{
\item{formula}{A \code{\link{formula}} object specifying the fixed and random
effects in standard R formula syntax. Random effects should be of the form
\code{(expr | factor)}.}

\item{data}{A \code{data.frame} containing all variables used in \code{formula}.}

\item{weights}{Optional. Column name, as in \code{\link{glmmTMB}} in \code{data} specifying
observation weights. Default is \code{NULL}, in which case all weights are 1.}

\item{link}{Character. Link function to use for the binomial response. Default is
\code{"logit"}.}

\item{penOpt}{List of penalty options, created by \code{\link{AUGglmmTMBPenalty}}:
\describe{
  \item{cfe}{Numeric. Strength of penalty on fixed effects. If \code{NULL},
    internally computed via \code{mv_multiplier()}. Setting \code{cfe = 0} disables fixed-effect penalty.}
  \item{autrepen}{Logical. If \code{TRUE}, automatically estimates the random-effects
    penalty parameters \eqn{\tau} and \eqn{\Psi} using the procedure described in Košuta et al (see also \code{\link{get_psi}}); the penalty is only applied to the 1st random effect.}
  \item{nu}{Optional list specifying random-effects penalty parameters. Can be of the same length as the number of specified random effects in which case the penalty is applied to all random effects; when shorter, the penalty is applied only to the first \code{length(nu)} random effects.}
  \item{psi}{Optional list of random-effects penalty matrices; need to be of the same length as \code{nu}.}
}}

\item{control}{List of control parameters created via \code{\link{AUGglmmTMBControl}}.}
}
\value{
A list with elements:
\describe{
  \item{fit}{The fitted penalized GLMM object, see also \code{\link{mpl_fitter}}.}
  \item{optre}{List with elements \code{opt_tau} and \code{opt_psi} corresponding
    to estimated random-effects penalty parameters if \code{autrepen = TRUE}, otherwise \code{NULL}.}
}
}
\description{
Fits a penalized binomial generalized linear mixed model with optional
penalties on both fixed and random effects using the approach of Košuta et al.
}
\details{
If \code{nu=NULL}, no random-effects penalty is applied and only the fixed-effects penalty is used.
If provided, \code{nu} and \code{psi} have to be lists, that can be of the same length as the number of specified random effects, in which case
the penalty is applied to all random effects. If shorter, the penalty is applied
only to the first \code{length(nu)} random-effects terms.
Setting \code{autrepen=TRUE} uses the data-driven procedure proposed by Košuta et al. to determine the penalty parameters; the parameter \code{nu} is set to \eqn{2q-1} internally, any other value supplied in \code{nu} is ignored.
}
\examples{
\dontrun{
data(birds)
fit <- AUGglmmTMB(
  cbind(parasites, 1 - parasites) ~ migration + (1 | species),
  data = birds,
  weights = NULL,
  link = "logit",
  penOpt = AUGglmmTMBPenalty(cfe = 1),
  control = AUGglmmTMBControl(fit_pGLM = TRUE, maxiter = 1)
)
summary(fit$fit)
}

}
\seealso{
\code{\link{AUGglmmTMBPenalty}},\code{\link{AUGglmmTMBControl}}, \code{\link{get_psi}}, \code{\link{mpl_fitter}}
}
