% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{AUGglmmTMBControl}
\alias{AUGglmmTMBControl}
\title{Control options for AUGglmmTMB fitting}
\usage{
AUGglmmTMBControl(
  fit_pGLM = FALSE,
  method_pGLM = "brglm2",
  maxiter = 50,
  tol = 1e-06,
  save_coef = FALSE,
  inter_iter = 1000,
  use_previous = FALSE
)
}
\arguments{
\item{fit_pGLM}{Logical. If \code{TRUE}, uses the penalized GLM step (Algorithm 2
in Košuta et al.), otherwise it uses Algorithm 1. Default is \code{FALSE}.}

\item{method_pGLM}{Character. The method to be used when fitting penalized GLM. Currently supported values are \code{"brglm2"} and \code{"logistf"}. \code{"logistf"} can only be used for Bernoulli outcomes when \code{link_fun="logit"}. Ignored when \code{fit_pGLM=FALSE}. Defaults to \code{"brglm2"}.}

\item{maxiter}{Integer. Maximum number of iterations for the penalized likelihood
fitting procedure. Default is 50.}

\item{tol}{Numeric. Convergence tolerance based on changes in parameter estimates.
Default is 1e-6.}

\item{save_coef}{Logical. If \code{TRUE}, stores coefficient estimates at each
iteration. Default is \code{FALSE}.}

\item{inter_iter}{Integer. Number of internal iterations used by
\pkg{glmmTMB}. Default is 1000.}

\item{use_previous}{Logical. If \code{TRUE}, uses estimates from the previous
iteration as starting values. Default is \code{FALSE}.}
}
\value{
A named list containing all specified control parameters.
}
\description{
Create a list of control parameters to customize the fitting behavior of
\code{\link{AUGglmmTMB}}.
}
\examples{

#Algorithm 1 from Košuta et al.
ctrl <- AUGglmmTMBControl(fit_pGLM = FALSE)

#Algorithm 2 from Košuta et al.
ctrl2 <- AUGglmmTMBControl(fit_pGLM = TRUE)

#single-iteration approximation to MPL from Košuta et al.
ctrl3 <- AUGglmmTMBControl(fit_pGLM = TRUE,maxiter=1)

}
