% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{make_pd_fix}
\alias{make_pd_fix}
\title{Construct augmented data for fixed-effects penalization}
\usage{
make_pd_fix(data, cfe, fix_beta, link_fun)
}
\arguments{
\item{data}{A list containing the model data with elements \code{Y}, \code{M},
\code{X}, \code{W}, and random-effects components \code{Z1, Z2, ...}
with corresponding grouping variables \code{grouping1, grouping2, ...}.}

\item{cfe}{Numeric. Strength of the fixed-effects penalty.}

\item{fix_beta}{Numeric vector of fixed-effects coefficients used to compute
the linear predictor.}

\item{link_fun}{Character. Link function. Supported options are
\code{"logit"}, \code{"probit"}, \code{"cloglog"}, \code{"cauchit"},
and \code{"loglog"}.}
}
\value{
A list with the same structure as \code{data}, containing the augmented
response \code{Y}, total counts \code{M}, design matrix \code{X}, weights
\code{W}, and corresponding random-effects components. The augmented dataset
includes two additional pseudo-observations per original observation.
}
\description{
Creates an augmented dataset by adding pseudo-observations corresponding to
the fixed-effects penalty.
}
\details{
The function constructs pseudo-data corresponding to the fixed-effects penalty
using a data-augmentation approach.
}
\examples{
data(birds)
Y<-birds$parasites
X <- model.matrix(parasites~migration+food, birds)
Z1<-model.matrix(parasites~migration, birds)
Z2<-model.matrix(parasites~1, birds)
grouping1<-as.numeric(as.factor(birds$Phylogenetic.Tomi))
grouping2<-as.numeric(as.factor(birds$species))
xdf<-list(Y=Y,X=X,Z=list(Z1,Z2),
  grouping=list(grouping1,grouping2),
  M=rep(1,nrow(X)),W=rep(1,nrow(X)))

xdft<-AUGglmmTMB:::restructure_data(xdf)
pd_fix<-make_pd_fix(xdft,mv_multiplier(xdft),rep(0,ncol(xdft$X)),"logit")

lapply(xdft,function(x) if(is.null(dim(x))) length(x) else dim(x) )
lapply(pd_fix,function(x) if(is.null(dim(x))) length(x) else dim(x) )

}
