% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{mv_multiplier}
\alias{mv_multiplier}
\title{Compute fixed-effects penalty strength}
\usage{
mv_multiplier(data)
}
\arguments{
\item{data}{A list containing the model data. Must include the element
\code{X} for the fixed effects design matrix. Optionally it also includes \code{M} for the number of trials and  \code{W} for the weights; if missing they are set to 1 internally for all observations.}
}
\value{
A numeric value equal to \eqn{\sqrt{p/n}}, where \eqn{p} is the number
of fixed-effect parameters and \eqn{n} is the effective sample size.
}
\description{
Computes the penalty strength for the fixed effects as suggested by Košuta et al.
}
\details{
The penalty strength is defined as \eqn{\sqrt{p/n}}, following the recommendation
of Košuta et al., where \eqn{p} is the number of columns of \code{X} and \eqn{n}
is the effective sample size.
}
\examples{
data(birds)
Y<-birds$parasites
X <- model.matrix(parasites~migration+food, birds)
Z1<-model.matrix(parasites~migration, birds)
Z2<-model.matrix(parasites~1, birds)
grouping1<-as.numeric(as.factor(birds$Phylogenetic.Tomi))
grouping2<-as.numeric(as.factor(birds$species))
xdf<-list(Y=Y,X=X,Z=list(Z1,Z2),
  grouping=list(grouping1,grouping2),
  M=rep(1,nrow(X)),W=rep(1,nrow(X)))

mv_multiplier(xdf)

}
